(wip) linux support)

+++ examples/src/res.cpp
@@ -1,5 +1,6 @@
 #include "res.h"
 #include "PopWork/resources/resourcemanager.h"
+#include <cstdint>
 
 using namespace PopWork;
 
@@ -298,7 +299,7 @@ static PopWork::ResourceId GetIdByVariable(const void *theVariable)
 			aMap[*(int *)gResources[i]] = i;
 	}
 
-	MyMap::iterator anItr = aMap.find((int)theVariable);
+	MyMap::iterator anItr = aMap.find((uintptr_t)theVariable);
 	if (anItr == aMap.end())
 		return RESOURCE_ID_MAX;
 	else
